requirements.txt
numpy
matplotlib
Pillow
scikit-learn
torch
torchvision
torchaudio
scipy
albumentations
tqdm
IPython